+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to
 - ✨(backend) allow use to stop conversation streaming #14
 - 🐛(agent) add the current date in the system prompt #18
 - ✨(backend) add feature flags from posthog #13
+- ✨(user) allow to use conversation data for analytics #23
 
 
 [unreleased]: https://github.com/numerique-gouv/conversations/compare/HEAD...main